Quizard_server/requirements.txt
 flask
 requests
 pandas
 google-generativeai
 flask-cors